serialization and error handler

+++ src/OtherNodes.hpp
@@ -61,7 +61,7 @@ public:
     void check() override;
 
     [[nodiscard]] inline std::vector<ExprType> getExprTypes() const { return m_exprTypes; }
-    [[nodiscard]] inline int getSize() const { return m_exprTypes.size(); }
+    [[nodiscard]] inline size_t getSize() const { return m_exprTypes.size(); }
     [[nodiscard]] inline ExprType getType(size_t i) const {
         DebugAssert(i >= m_exprTypes.size(), "Index out of range");
         return m_exprTypes[i];
